OBPIH-5433 Fixes after review
