+++ grails-app/controllers/org/pih/warehouse/inventory/StockMovementController.groovy
@@ -12,11 +12,9 @@ package org.pih.warehouse.inventory
 
 import grails.converters.JSON
 import org.grails.plugins.csv.CSVWriter
-import org.pih.warehouse.api.NotificationType
 import org.pih.warehouse.api.StockMovement
 import org.pih.warehouse.api.StockMovementDirection
 import org.pih.warehouse.api.StockMovementItem
-import org.pih.warehouse.api.StockMovementType
 import org.pih.warehouse.core.ActivityCode
 import org.pih.warehouse.core.BulkDocumentCommand
 import org.pih.warehouse.core.Constants
@@ -26,9 +24,7 @@ import org.pih.warehouse.core.DocumentType
 import org.pih.warehouse.core.Location
 import org.pih.warehouse.importer.ImportDataCommand
 import org.pih.warehouse.order.Order
-import org.pih.warehouse.order.OrderTypeCode
 import org.pih.warehouse.picklist.PicklistItem
-import org.pih.warehouse.requisition.Requisition
 import org.pih.warehouse.requisition.RequisitionSourceType
 import org.pih.warehouse.requisition.RequisitionStatus
 import org.pih.warehouse.shipping.Shipment
@@ -211,11 +207,10 @@ class StockMovementController {
             if (stockMovement?.shipment?.currentStatus == ShipmentStatusCode.PENDING || !stockMovement?.shipment?.currentStatus) {
                 try {
                     stockMovementService.deleteStockMovement(params.id)
-                    params.notificationMessage = g.message(
+                    flash.message = g.message(
                             code: 'react.stockMovement.deleted.success.message.label',
                             body: 'Stock Movement has been deleted successfully',
                     )
-                    params.notificationType = NotificationType.SUCCESS
                 } catch (Exception e) {
                     log.error("Unable to delete stock movement with ID ${params.id}: " + e.message, e)
                     flash.message = "Unable to delete stock movement with ID ${params.id}: " + e.message
@@ -245,7 +240,7 @@ class StockMovementController {
                 (currentLocation == stockMovement.destination) ? StockMovementDirection.INBOUND : "ALL"
 
 
-        redirect(action: "list", params: params)
+        redirect(action: "list", params: params << ["flash": flash as JSON])
     }
 
     def requisition = {
