--- src/groovy/org/pih/warehouse/api/NotificationType.groovy
@@ -1,14 +1,0 @@
-package org.pih.warehouse.api
-
-enum NotificationType {
-    SUCCESS('success'),
-    WARNING('warning'),
-    INFO('info'),
-    ERROR('error'),
-    ERROR_OUTLINED('error-outlined'),
-    ERROR_FILLED('error-filled')
-
-    String notificationType
-
-    NotificationType(String notificationType) { this.notificationType = notificationType }
-}
