+++ src/js/hooks/useFlashScopeListener.jsx
@@ -1,0 +1,36 @@
+import { useEffect } from 'react';
+
+import _ from 'lodash';
+import queryString from 'query-string';
+import { useHistory, useLocation } from 'react-router-dom';
+
+import notification from 'components/Layout/notifications/notification';
+import NotificationType from 'consts/notificationTypes';
+
+const useFlashScopeListener = () => {
+  const { search } = useLocation();
+  const history = useHistory();
+  const parsedQuerySearch = queryString.parse(search);
+  const { flash } = parsedQuerySearch;
+
+  const clearQuerySearch = (...pathToSkip) => {
+    const clearedQuerySearch = _.omit(parsedQuerySearch, pathToSkip);
+    const parsedClearedQuerySearch = queryString.stringify(clearedQuerySearch);
+    history.replace({
+      search: parsedClearedQuerySearch,
+    });
+  };
+
+  useEffect(() => {
+    if (!flash) {
+      return;
+    }
+    const { message } = JSON.parse(flash);
+    if (message) {
+      notification(NotificationType.SUCCESS)({ message });
+    }
+    clearQuerySearch('flash');
+  }, []);
+};
+
+export default useFlashScopeListener;
