+++ src/js/wrappers/FlashScopeListenerWrapper.jsx
@@ -1,0 +1,9 @@
+import useFlashScopeListener from 'hooks/useFlashScopeListener';
+
+const FlashScopeListenerWrapper = ({ children }) => {
+  useFlashScopeListener();
+
+  return children;
+};
+
+export default FlashScopeListenerWrapper;
