--- src/js/wrappers/NotificationListenerWrapper.jsx
@@ -1,9 +1,0 @@
-import useNotificationListener from 'hooks/useNotificationListener';
-
-const NotificationListenerWrapper = ({ children }) => {
-  useNotificationListener();
-
-  return children;
-};
-
-export default NotificationListenerWrapper;
